#397 refactor: create tree controller
