+++ src/Controller/AppController.php
@@ -12,7 +12,6 @@
  */
 namespace App\Controller;
 
-use BEdita\SDK\BEditaClientException;
 use BEdita\WebTools\ApiClientProvider;
 use Cake\Controller\Controller;
 use Cake\Core\Configure;
@@ -373,41 +372,5 @@ class AppController extends Controller
         $objectNavModule = (string)$session->read('objectNavModule');
 
         return (array)Hash::get($objectNav, sprintf('%s.%s', $objectNavModule, $id), []);
-    }
-
-    /**
-     * Load tree data.
-     *
-     * @return void
-     */
-    public function treeJson(): void
-    {
-        $this->request->allowMethod(['get']);
-        $query = $this->Modules->prepareQuery($this->request->getQueryParams());
-        $root = Hash::get($query, 'root');
-        $full = Hash::get($query, 'full');
-        unset($query['full']);
-        unset($query['root']);
-        try {
-            if (empty($full)) {
-                $key = (empty($root)) ? 'roots' : 'parent';
-                $val = (empty($root)) ? '' : $root;
-                if (empty($query['filter'])) {
-                    $query['filter'] = [];
-                }
-                $query['filter'][$key] = $val;
-            }
-            $response = $this->apiClient->getObjects('folders', $query);
-        } catch (BEditaClientException $error) {
-            $this->log($error, 'error');
-
-            $this->set(compact('error'));
-            $this->set('_serialize', ['error']);
-
-            return;
-        }
-
-        $this->set((array)$response);
-        $this->set('_serialize', array_keys($response));
     }
 }
